requirements.txt
waitress~=2.1.2
python-dotenv~=1.0.0
requests~=2.28.2
Flask~=2.2.3
aiohttp~=3.8.4
PyJWT~=2.6.0